requirements.txt
Flask==2.0.1
selenium==4.1.0
requests==2.26.0
python-dotenv==0.19.0
fake_useragent==2.0.3